// typechain/src/MarketPlace.ts
export declare namespace MarketPlace {
  export type ValidateTradeStruct = {
    tokenId: PromiseOrValue<BigNumberish>;
    collectionId: PromiseOrValue<BigNumberish>;
    owner: PromiseOrValue<string>;
  };

  export type ValidateTradeStructOutput = [BigNumber, BigNumber, string] & {
    tokenId: BigNumber;
    collectionId: BigNumber;
    owner: string;
  };

  export type TradeStruct = {
    uri: PromiseOrValue<string>;
    collectionId: PromiseOrValue<BigNumberish>;
    owner: PromiseOrValue<string>;
  };

  export type TradeStructOutput = [string, BigNumber, string] & {
    uri: string;
    collectionId: BigNumber;
    owner: string;
  };

  export type CardStruct = {
    id: PromiseOrValue<string>;
    collectionId: PromiseOrValue<BigNumberish>;
    acceptedCurrencies: PromiseOrValue<string>[];
    owner: PromiseOrValue<string>;
    done: PromiseOrValue<boolean>;
    acceptedCurrency: MarketPlace.TradeStruct;
    spotId: PromiseOrValue<BigNumberish>;
  };

  export type CardStructOutput = [
    string,
    BigNumber,
    string[],
    string,
    boolean,
    MarketPlace.TradeStructOutput,
    BigNumber
  ] & {
    id: string;
    collectionId: BigNumber;
    acceptedCurrencies: string[];
    owner: string;
    done: boolean;
    acceptedCurrency: MarketPlace.TradeStructOutput;
    spotId: BigNumber;
  };
}